update workflow configs
